se agrega un return si el parametro en null

+++ utils/handlerDeleteFile.js
@@ -2,6 +2,8 @@ const fs = require('fs')
 const path = require('path')
 
 module.exports.handlerDeleteFile = (pathFile) => {
+  if (!pathFile) return
+
   const cd = path.join(__dirname, '../')
   if (fs.existsSync(`${cd}${pathFile}`)) {
     fs.unlinkSync(`${cd}${pathFile}`)
